redesign of the application structure
