+++ src/main/java/com/janluk/schoolmanagementapp/common/model/SubjectEntity.java
@@ -1,6 +1,6 @@
-package com.janluk.schoolmanagementapp.model;
+package com.janluk.schoolmanagementapp.common.model;
 
-import com.janluk.schoolmanagementapp.model.vo.SubjectType;
+import com.janluk.schoolmanagementapp.common.model.vo.SubjectType;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
