--- src/main/java/com/janluk/schoolmanagementapp/repository/port/UserRepository.java
@@ -1,5 +1,0 @@
-package com.janluk.schoolmanagementapp.repository.port;
-
-public interface UserRepository {
-
-}
